Added rendering sub/superscripts
I also kept poking around with spacing during this time so a couple
other things related to that might have changed too.  I gave up on
trying to get make the official TeX spacing work and started tweaking
things myself manually.

+++ src/ganim/object/text/gex/noad.hpp
@@ -26,7 +26,7 @@ namespace ganim::gex {
         Vcent  // Produced by \vcenter
     };
     // Boolean says whether to add the spacing for script styles
-    std::pair<int, bool> get_atom_spacing(AtomType left, AtomType right);
+    std::pair<double, bool> get_atom_spacing(AtomType left, AtomType right);
 
 
     class Noad;
